Allow the filter global function
The Sass function shadows a CSS function. The global name should still
be allowed when using the CSS function.
A more complete change might be to allow only global usages that pass a
number as the only argument, but this is more complex to implement
(especially when the argument comes from a Sass variable that might be a
number or no).